requirements.txt
certifi==2019.3.9
cffi==1.12.0
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
greenlet==0.4.13
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
python-dotenv==0.10.1
readline==6.2.4.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
urllib3==1.24.1
Werkzeug==0.15.1
